ml_vision/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# SourceGraph Vision Node — Python dependencies
#
# Install order matters for CUDA builds:
#   1. Install PyTorch with the correct CUDA wheel FIRST (see note below).
#   2. Then install everything else.
#
# ── PyTorch / CUDA ────────────────────────────────────────────────────────────
# Do NOT install torch/torchvision from PyPI plain — they ship CPU-only wheels.
# Use the official CUDA index instead.  For CUDA 12.x (RTX 3050 / Ampere):
#
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# The lines below are kept for requirements.txt completeness and are safe to
# install after the CUDA wheel is already in the environment.
# ─────────────────────────────────────────────────────────────────────────────

# Core ML
torch>=2.2.0
torchvision>=0.17.0

# HuggingFace — CLIP model + fast tokeniser
transformers>=4.40.0
# Accelerates HF tokeniser; optional but strongly recommended
tokenizers>=0.19.0

# YOLOv8
ultralytics>=8.2.0

# Image I/O
Pillow>=10.3.0

# FastAPI stack
fastapi>=0.111.0
uvicorn[standard]>=0.29.0       # Standard = includes uvloop + httptools
python-multipart>=0.0.9          # Required for UploadFile / form parsing

# Async HTTP client (Orchestrator bounce)
httpx>=0.27.0

# Numerical
numpy>=1.26.0

# Optional but useful in production
pydantic>=2.7.0                  # Already a FastAPI dependency; explicit pin